refactor: introduce types index
