+++ src/components/SupportCardList/SupportCard/index.ts
@@ -1,5 +1,3 @@
 import SupportCardComponent from "./SupportCard";
-import SupportCard from "./SupportCard";
 
-export { SupportCard }
 export default SupportCardComponent
